api/requirements.txt
Flask
requests
smbus2
python-dotenv